# prepare perf: make ProgramBuilder aware of plan to count/estimate required memory